prueba_docker/requirements.txt
#mapfre_agentkit @ git+https://${GIT_TOKEN}@github.com/mapfre-tech/esp-ateneaes-mapfre-agent-kit@feature/multimodal
pyyaml>=6.0

